Improve binary prediction logic for cosmovisor
Some of the nodes might want to switch to cosmovisor on nodes that
already have some state, but if they sync one or more upgrades genesis
bin will corrupt chain date. Due to the file upgrade-info.json store
meta information for upgrade it can be used to do better decision.